requirements.txt
streamlit==1.36.0
opencv-python-headless==4.7.0.72
pillow==10.3.0
numpy==1.25.2
deepface==0.0.92
ml-dtypes==0.2.0
tensorflow==2.15.0
keras==2.15.0
